+++ src/ipm/ipx/sparse_matrix.cc
@@ -195,13 +195,13 @@ void MultiplyAdd(const SparseMatrix& A, const Vector& rhs, double alpha,
     const Int m = A.rows();
     const Int n = A.cols();
     if (trans == 't' || trans == 'T') {
-        assert(rhs.size() == m);
-        assert(lhs.size() == n);
+        assert((Int)rhs.size() == m);
+        assert((Int)lhs.size() == n);
         for (Int j = 0; j < n; j++)
             lhs[j] += alpha * DotColumn(A, j, rhs);
     } else {
-        assert(rhs.size() == n);
-        assert(lhs.size() == m);
+        assert((Int)rhs.size() == n);
+        assert((Int)lhs.size() == m);
         for (Int j = 0; j < n; j++)
             ScatterColumn(A, j, alpha*rhs[j], lhs);
     }
@@ -211,8 +211,8 @@ void AddNormalProduct(const SparseMatrix& A, const double* D, const Vector& rhs,
                       Vector& lhs) {
     const Int m = A.rows();
     const Int n = A.cols();
-    assert(rhs.size() == m);
-    assert(lhs.size() == m);
+    assert((Int)rhs.size() == m);
+    assert((Int)lhs.size() == m);
     for (Int j = 0; j < n; j++) {
         double temp = DotColumn(A, j, rhs);
         if (D) temp *= D[j]*D[j];
